+++ src/wht/rudderPB.py
@@ -106,3 +106,33 @@ class RudderPB:
                 f"Could not find material file prefix {material_file_prefix} in the output of pb compile command: {pb_compile_output}"
             )
         return model_hash, features_profiles_model
+
+    def split_material_table(material_table_name: str) -> Tuple:
+        """
+        Splits given material table into model_name, model_hash and seq_no
+        Ex. Splits "Material_user_var_table_54ddc22a_383" into (user_var_table, 54ddc22a, 383)
+
+        Args:
+            material_table_name: material table name
+        Returns:
+            Tuple: returns ("model_name", "model_hash", seq_no)
+        """
+        mlower = material_table_name.lower()
+        if MATERIAL_PREFIX not in mlower:
+            logger.warning(
+                f"Couldn't split {material_table_name}, it does not contain table prefix '{MATERIAL_PREFIX}'"
+            )
+            return (None, None, None)
+
+        try:
+            table_suffix = mlower.split(MATERIAL_PREFIX)[-1]
+            split_parts = table_suffix.split("_")
+            seq_no = int(split_parts[-1])
+            model_hash = split_parts[-2]
+            model_name = "_".join(split_parts[0:-2])
+            return (model_name, model_hash, seq_no)
+        except (IndexError, ValueError):
+            logger.warning(
+                f"Couldn't split the material table {material_table_name} into model name, hash, and seq_no"
+            )
+            return (None, None, None)
